finished the based part

+++ store-dev/store-service/src/main/java/service/impl/UserServiceImpl.java
@@ -1,5 +1,10 @@
 package service.impl;
 
+import exception.tx.AddContactInfoTxException;
+import exception.tx.AddOrderInfoTxException;
+import exception.tx.AddOrdersTimeTxException;
+import exception.tx.AddOrdersTxException;
+import exception.tx.DeleteCartInfoTxException;
 import exception.tx.TxException;
 import exception.tx.AddCartTxException;
 import exception.tx.AddCartInfoTxException;
@@ -15,6 +20,8 @@ import pojo.CartCollection;
 import pojo.CartInfo;
 import pojo.CartItem;
 import pojo.ContactInfo;
+import pojo.Orders;
+import pojo.OrdersTime;
 import pojo.User;
 import service.UserService;
 
@@ -83,7 +90,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<CartItem> getCartItems(int userId) {
         List<Integer> cartInfoIds = userMapper.getCartInfoIdByUserId(userId);
-        return userMapper.getCartInfo(cartInfoIds);
+        return cartInfoIds.size() == 0 ? null : userMapper.getCartInfo(cartInfoIds);
     }
 
     @Override
@@ -101,19 +108,53 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public int submitOrder(int userId, ContactInfo contactInfo) {
-        //增加联系人信息，返回id
-        int contactInfoId = userMapper.addContactInfo(contactInfo);
-        //根据用户id查询所有的cartInfoId->getCartInfoIdByUserId
-        List<Integer> cartInfoIds = userMapper.getCartInfoIdByUserId(userId);
-        //根据得到的本用户的所有cartInfoId查询对应的cartItem->getCartInfo
-        List<CartItem> cartItems = userMapper.getCartInfo(cartInfoIds);
-        //将cartInfo表的记录批量删除
-        int rowsForDelete = userMapper.deleteCartInfos(cartItems);
-        //将得到的cartItems批量插入orderInfo表得到orderInfoId的集合
-        int rowsForOrderInfo = userMapper.addOrderInfo(cartItems);
-        //操作orders表，其插入行数应为orderInfoId的size大小
-        int rowsForOrders = userMapper.addOrders(userId, contactInfoId, cartItems);
-        return 0;
+    @Transactional
+    public int submitOrder(int userId, ContactInfo contactInfo) throws TxException {
+        DefaultTransactionDefinition defaultTransactionDefinition = new DefaultTransactionDefinition();
+        TransactionStatus transactionStatus = dataSourceTransactionManager.getTransaction(defaultTransactionDefinition);
+        try {
+            //增加orderTime
+            OrdersTime ordersTime = new OrdersTime();
+            int rowForTime = userMapper.addOrdersTime(ordersTime);
+            if (rowForTime != 1) {
+                throw new AddOrdersTimeTxException("添加订单时间失败");
+            }
+            //增加联系人信息，取得id
+            int rowForContact = userMapper.addContactInfo(contactInfo);
+            if (rowForContact != 1) {
+                throw new AddContactInfoTxException("添加收货人信息失败");
+            }
+            //根据用户id查询所有的cartInfoId
+            List<Integer> cartInfoIds = userMapper.getCartInfoIdByUserId(userId);
+            //根据得到的本用户的所有cartInfoId查询对应的cartItem
+            List<CartItem> cartItems = userMapper.getCartInfo(cartInfoIds);
+            //将cartInfo表的记录批量删除
+            int rowsForDelete = userMapper.deleteCartInfos(cartItems);
+            if (rowsForDelete != cartItems.size()) {
+                throw new DeleteCartInfoTxException("删除cartItems有误");
+            }
+            //将得到的cartItems批量插入orderInfo表得到orderInfoId的集合
+            int rowsForOrderInfo = userMapper.addOrderInfo(cartItems);
+            if (rowsForOrderInfo != cartItems.size()) {
+                throw new AddOrderInfoTxException("添加orderInfos有误");
+            }
+            //操作orders表，其插入行数应为orderInfoId的size大小
+            int rowsForOrders = userMapper.addOrders(userId, contactInfo.getId(), ordersTime.getId(), cartItems);
+            if (rowsForOrders != cartItems.size()) {
+                throw new AddOrdersTxException("添加orders有误");
+            }
+            dataSourceTransactionManager.commit(transactionStatus);
+            return rowsForOrders;
+        } catch (AddOrdersTimeTxException | AddContactInfoTxException | DeleteCartInfoTxException | AddOrderInfoTxException | AddOrdersTxException e) {
+            System.out.println(e.getMessage());
+            dataSourceTransactionManager.rollback(transactionStatus);
+            throw new TxException(e.getMessage());
+        }
+    }
+
+    @Override
+    public List<Orders> getOrders(int userId) {
+        List<Orders> orders = userMapper.getOrders(userId);
+        return orders.size() == 0 ? null : orders;
     }
 }
